Fix reading error for bundle carrier-config items
Carrier configuration items of bundle type were adjusted by ImsStack
when the configuration is successfully loaded. But, it was not set
properly while processing the value of each item because the count value
of the adjusted configuration item was not set.
This is a change to correct the count value when each configuration item
is moved to the bundled configuration.

Bug: 229558259
Test: manual (verify that bundle configuration is read)
Change-Id: I2717e3fc4867b066fdbfab79670986b6dd04b757