Increase modified time of configuration xml
The trace option values were changed recently. When installing the
recent ImsStack apk, the log is disabled because the trace option is
matched between the configuration xml and source code.
So, this change updates the version of the configuration xml to
re-create the configuration database when installing the recent ImsStack
apk.

Test: m(build)
Change-Id: I15b63a5c479669ec6bb59513beef405a2510666e