Preserve SIP message handling order
Basically, SIP message is handled by sipcore engine in order, but when
it almost receives the SIP request and response at the same time, the
SIP response is always handled first because SIP request is handled
after posting the message internally. To preserve the message handling
order, SIP client connection will handle the response after posting the
message rather than handling in the flow of the listener's method call.

Test: manual (ims-reg verified on cuttlefish)

Change-Id: Ied92fd263fe3aa454bff47186b0c97eeebdccf1f