(Ut) Implementation to handle N99 HTTP error of carrier config
This is an implementation to handle N99 error codes from carrier
configuration value. When N99 is set as a block codes for HTTP error,
all of Nxx error codes should be blocked.
e.g) If carrier config has a value 499, then all of 4xx error cases
handled as blocking codes.

Test: Unit test
Change-Id: I0cc8f8b8417cc6ec29fedfd4fc8798842bbbc03b