(Mtc) make the JNI interface from the java layer when incoming call received.
In the previous CL ag/18195538, the Native just notify(pre-incoming) the java layer there's a incoming call with expecting instance of MtcCall.java will be made at the proper timing and then immediately process remain works to get incoming call using another thread.
In this case, can not guarantee instance of MtcCall.java ready to take
event through the JNI interface from the Native at a specific point.
so, To address this problem, need to make the JNI interface from the java layer with MT case unlike before(it was made by calling from the Native) and go further with it.

Test: mm build and DISS MO/MT
Change-Id: Idf47527129f2d30cd884bb6d18b68f79fe7e3aa8