Fix a failure to load carrier configuration
IMS registration procedure was not conducted when the device boots
without SIM card and inserting the SIM card after boot completed. This
situation was caused by a failure to load a carrier configuration in the
native layer when inserting the SIM card.
This problem is fixed by resetting the system call interface which
provides an interface to get a carrier configuration from Java layer
when restarting the IMS services.

Bug: 232815134
Test: manual (boot the device w/o SIM card and insert SIM card)
Change-Id: Ic6a9449929f888fc9511805ce44ab23e5faaad76