Rewrite call control by USIM function
This commit rewrites the call control by USIM function based on new USAT
interface and adds more logics to handle the various cases.
If the USIM changes the call control type or dialed string or media
type, the ImsStack handles the ongoing call according to the modified
information if it is capable of the operations, but if not, the ImsStack
will notify the call failure with a proper reason code.

Bug: 227272715
Test: manual (verify call control by USIM)
Change-Id: I37161ab1ef21014a9771f0bece757ad133d40128