Remove or replace ImsService's registry
This commit is to remove the unused code and replace the ImsService's
registry to new one in the ImsService implementation layer.
All the call and registration states are managed and controlled by the
native Aos and Mtc enablers. And, IMS on/off state is moved into new
registry and others are removed.

Test: manual (verify IMS on/off change notification)
Change-Id: I8364ec2430e4fd2b0f90380b74b3ef2ad1a37cc5