Fix a conflict of the ISIM records' refresh
This commit is to fix not reading the ISIM records when ISIM refresh
commands are requested twice for a short time. Even though the refresh
is in progress by the carrier configuration change, if the refresh by
the enabler's restart is requested, then the last request will be
handled by the subscriber's configuration by ignoring the previous
request.

Bug: 237107221
Test: manual (verify ims-reg when removing & inserting SIM card)
Change-Id: If876a9b7fe2ffb675716508515621de5b7daa210