Fix SIP server connection handling on TCP
This commit is to fix SIP server connection handling when it's created
using the TCP transport protocol. Basically, the SIP server transaction
that uses the TCP transport protocol is transited to the TERMINATED
state immediately after sending the SIP response.
So, if a new TCP connection needs to be established when sending the
response message, it needs more time to establish the TCP connection,
but this response isn't sent to the network because the SIP server
connection is also closed immediately after the send operation.

This patch contains the followings:
  1) Wait for sending SIP response and close the SIP server connection
  2) Re-use the TCP socket if the destination address is same
     (Preventing to create a new TCP socket when sending SIP messages)

Bug: 234416713
Bug: 237482426
Bug: 237740102
Bug: 237740533
Test: manual (verify ims-reg/call on target device)
Test: manual (verify Verizon TC - IMSVoIP_02.04, Voip 3.29.1)
Change-Id: I9c2e9dc926426d4fecf7da205f3215fb40792875