(Mtc) Add config asset: message_type_support_geolocation_pidf_int_array
A config is redefined (no code change for this):
GEOLOCATION_PIDF_IN_SIP_INVITE_SUPPORT_INT_ARRAY is for the session
initiation, not limited to INVITE.

The new config determines which message needs PIDF-LO if the above
config is on in order to meet AT&T requirement.
The array can contains the following values.

- INVITE = 0,
- PROVISIONAL_RESPONSE = 1,
- FINAL_SUCCESS_RESPONSE = 2,
- FINAL_FAILURE_RESPONSE = 3,

Bug: 237051706
Test: Turn on the config, make MO/MT call
Change-Id: I8042e0e0b7689cc7eb764cc48568ac916932ec15