Load carrier-config explicitly when restarting enablers
In the current implementation, the carrier-config in the native layer
will be updated by the change notification of the carrier-config.
However, while restarting enablers in the multi-sim environment, this
refresh procedure was delayed and the carrier-config in the native layer
was not updated before refreshing by the internal configuration
components. So, newly started enablers used the old configuration.
To guarantee that the carrier-configs are updated before refreshing, if
the enablers' are newly created, the carrier-configs will be updated
explicitly for each SIM slot.
And, this also contains not to set the listener to receive the change
notification of the carrier-config in the common configuration
components.

Bug: 238157147
Test: manual (verify ims-reg while SIM hot swap)
Change-Id: I111ed0a459a9cc027a3f0250ab752dec24ee04b2