Add a config to handle parameters in P-ANI header
This commit is to add a conifguration to handle parameters in P-ANI
header. Some of carriers request to use all ZEROs as the MAC address in
P-ANI header, and some request to add the country parameter in P-ANI
header. To handle these requirements, this CL adds two more internal
configuration items and redeems its related logics.

The following asset configuration items are added:
 - hide_mac_address_in_pani_header_bool
 - support_country_param_in_pani_header_bool

Bug: 238829608
Test: manual (verify P-ANI header on ims-reg)
Test: atest ImsStackNativeEngineTests
Change-Id: I47828937801c2f31bc4d5aa7002f3480a0c56713