Fix wrong Allow header in REGISTER request
SipConfig loads the configurations to set Allow header for each SIP
method. When the configuration is refreshed, the member variable for the
string array that will be filled in the Allow header is not initialized,
so the same method name was added multiple times whenever the refresh is
triggered. To avoid this situation, the initialization of the member
variable is added.

Bug: 240105138
Test: atest ImsStackNativeConfigTests
Test: manual (verify ims-reg on target device)
Change-Id: Ie7077124c271c0079b7ca95d88246a028efa07d6