(Mtc) reflect ISession interface change about refreshing
This commit is to use new API in ISession.
As-Is : MtcCall has a member variable m_bRefreshing
this is set by IRefreshListener#Refresh_NotifyTimerExpired()
and reset by #Refresh_NotifyCompleted()
To-do : MtcCall checks ISession#IsSessionRefreshInProgress()

Remark: after #Refresh_NotifyCompleted() is called, the new
api still returns true so Mtc needs to post a message.

Test: atest passes, verified using simulator
Change-Id: I916f9f138f358409d9648f19545764cc3f813be4