Improve SIP message routing with callee preference
If the enabler sets the filter criteria with the callee preference, it
needs to be handled with priority among the same filter criterias
without the callee preference. In some cases, it didn't work.
So the logic is changed to give a high score to determine the routing of
the SIP message if the callee preference is set.

Bug: 241863227
Test: manual (verify the SIP message routing w/ callee preference)
Change-Id: Ica37d97d9fea4e3b2cb6cbc344c6e24d83ff70e3