Set the AppName to the full name of activity
I set the AppName to UceApp rather than full name of the activity. And set as a AppName in PublishManager.
PublishManager sends a message using the AppName.
However, it`s not delivered because AppName is not full name.

So I change the AppName to the full name of Uce Activity.

Bug: b/242962435
Test: Publish - 200OK - send a message to the UceApp in TMobile live network

Change-Id: I7ea6abab0a636067a8df865541deddc89717ef4c