(Mtc) Fix to follow RBT requirement of VZW
Verzion requires to send the RTP/RTCP packet even though UE plays
locally generated tone in the early dialog. So it has the change to run
the media session regardless of playing the local tone.
And every response or request in the early dialog state will be checked
to play network tone or local tone. UE will play the local ringback tone
only in the case of receiving 180 response.

Bug: b/257361057, b/237737336
Test: script test, manual test with DISS
Change-Id: I341245dab2e72527d688101f5908d4bb2f5d0e46