Added the no response case in publishing state.
If the publish is failed, GetPreviousResponse of IServiceMethod is null.
However, in this case, I deal with the engine is not working. So it sends a command error to UceAgent.

So I changed this case to handle the no response.
Added code to check if retry or exponential timer should be run.

Bug: b/246664967
Test: PUBLISH-200OK-REFRESH-200OK in TMO live network
Change-Id: I74846feaf58f6a820ae96db52c5ca1a7a930b71d