Manage INetworkIpSec instance based on slot
INetworkIpSec instance was managed as a single object when ImsStack uses
the IpSec operations that are provided by Linux kernel. At this time,
we didn't need to manage the member variables with the thread-safe
operations. However, ImsStack currently uses AOSP APIs for IpSec
operations, and needs to manage the member varialbes with the
thread-safe operations. To achieve this purpose, ImsStack will manage
the INetworkIpSec instance separately based on the slot instead of the
thread-safe operations (using mutex).

Bug: 254749868
Test: manual (verify ims-reg and airplane mode on/off)
Test: atest ImsStackNativeTests
Change-Id: Ic372a25ff53bcbfb62e4ed40da29bd900039b1e0