Fix a crash when service is restarted
ImsStack restarts the services when the SIM card is changed or hot
swapped on runtime. At this point, ImsStack was crashed by accessing the
invalid memory that is already released. The crash was happened when
using the IMessageCallback between different threads.
ImsStack can avoid this crash so that if an instance uses an
IMessageCallback between threads, the instance will remove the pending
messages of other thread before that is being destroyed. And, it can
release its allocated memory for some parameters.

Bug: 251705564
Test: atest ImsStackNativePlatformTests
Test: manual (verify ims-reg/call, SIM remove/insert)
Change-Id: If2a175387f9b4ac540a698007f9220e6820e075a