Apply runtime logging options
This commit contains a change to apply the runtime logging options in
Java and native layer. A logging component in the end point of Java
layer will check the options and send a log when the log level is
enabled. This logging options will also be synchronized with the native
layer and the native layer also sends logs only when the specified log
level is enabled.

Bug: 246679092
Test: manual (verify that logs are displayed based on log level)
Test: atest ImsStackNativeConfigTests
Change-Id: I7179c988c9b3b964c9a0fc718c0f4d70accf1beb