(Mtc) fix issue - not support timer in Require
This commit is to fix an issue related 'timer' feature.
'timer' feature tag is added by SIP engine so it's not listed
in MTC Extension list. Because of this, MTC rejects an INVITE
including 'timer' feature in 'Require' header as a non-supported
extension. This commit adds 'timer' to MTC extension list but
doesn't control header value.

Bug: b/245033711
Test: atest passed and verified by DISS
Change-Id: I75ead229bf910de68e8d300dd295610837f1130c