(Aos) Modified to check actual RAT for updating "cs" feature tag
This CL is to fix VZW VoWiFi TC 3.49 failure.
AosHandle checked the mobile network type during the RAT is being
changed from LTE to NONE. (middle of the RAT_GUARD_TIME).
So, "cs" feature tag was not added into the Contact header.

This CL changes AosHandle to check actual RAT for updating "cs" feature
tag so that the feature tag is properly added.

Bug: 258130791
Test: gTest and TE
Change-Id: I1abaab6649b1b3fab64c2745ea74b12cedfbd194