Fix ISIM refresh when carrier-config changed
When the carrier-config is changed, ISIM records in SubscriberConfig are
clear, but it doesn't change the state of SubscriberConfig. So,
SubscriberConfig doesn't update the ISIM records even if it tried to
load the ISIM records. By this reason, IMS registration was not done
because the IMPU is not available.
This change is to proceed the ISIM refresh process when the
carrier-config is changed by changing the state properly.

Bug: 258386255
Test: manual (verify ims-reg on boot-up)
Test: manual (verify ims-reg after factory reset)
Change-Id: I8071d0b649360b2cd3cc89829bc45a2532238bc8