(Mtc) update EPS Fallback Wait timer
As per the result of the test in b/259333786,
the wait timer value for EPS Fallback (no network response case)
should be updated.
As-Is : same as mo_call_request_timeout (Timer_VZW : 6s)
To-Be : fixed value 12s.

If we need to add configuration to optimize the value, I'll add
later.

Test: atest passed
Change-Id: I0d9f1bfe0ace6943b4f40b1e4a8cca49f56f4df2