(Mtc) Change to use IMessageUtils directly instead of MessageUtil
MessageUtil is deprecated and all MTC module will use IMessageUtils.
This commit is change MessageFormatter class to do this and
the change also resolves an issue below
- when MessageFormatter::SetReasonHeader() is invoked by terminaiting
procedure, ImsStack process got crashed. It can occur if a call is
made in a slot other than slot_0 because MessageUtil uses current
thread's slot id but Terminate() is done by Java thread, not native
thread. Because of this, slot 0 is selected and there is NPE.

Bug: b/263326537
Test: atest passed
Change-Id: Id8cc2ff6af26831cd27bd00fb3391c063c8b3eb3