(Mtc) add a location based call value checking.
Some carriers require that latest location information when do emergency
call or wifi call or both of them with all number or designated
numbers. so from GII, adds a `SuppType::GEOLOCATION` value if location
based call is required at a this time call.
If a value is existed and true, that means latest location information has been updated. then, go with
this. but if a value is false, that means getting latest location information has been failed. then, no need to add geo location information.
Of course, don't need to care if there is a no `SuppType::GEOLOCATION` value.

Test: atest & build(mm)
Change-Id: I9f4209adc0f2a46afb66b7a9bdd0d28b14e05799