Change alarm timer API for device IDLE mode
Even when the device is in IDLE mode, the alarm timer in the IMS stack
must expire correctly to ensure the IMS registration state update. This
is an important factor in using phone features on the device.
So, this changes the alarm timer API from setExact to
setExactAndAllowWhileIdle in the IMS stack.

Bug: 258090040
Test: manual (verify ims-reg refresh while device is in IDLE mode)
Change-Id: I3ba62e8ef7e585ad1dc19a71f7585c7ea8fdac59