(Mtc) fix bug - crash when call is terminated by remote
Issue CL : ag/20892177
With above CL, MTC tries to invoke ISession#SetReasonHeaderSetter but
ISession is already deleted if BYE is received.
In the case, engine immediately transits to STATE_TERMINATED without
TERMINATNIG state because guaranteeing retransmission is not necessary
in that case. (SIP stack just aborts it) Because the state is
TERMINATED, MTC immediately destroys ISession and causes crash.
This commit is to release ISession after all MtcSession internal
operation is completed with the destructor.
This is a work-around solution and an architectural change might be
needed to cover this.

Bug: b/265083934
Test: verified MO / MT call and termninated by local / remote
Test: CANCEL received
Test: atest with hwasan
Change-Id: I0242f397d551388d8d11a0c53e978188218f0394