(Mtc) Bug fix - No auto accept if video port 0 but a=sendrecv
As media policy was changed to use the direction in the SDP even in case
the port number is 0, media notifies 'sendrecv', not 'invalid' in the
issue scenario in the bug below.
This commit is to change the condition of auto accept. If call type is
not changed and it's VoIP or RTT which doesn't contain video media, the
condition doesn't check the change of video media diration.

Bug: b/264975195
Test: atest with hwasan
Test: video port 0 and sendrecv scenario verified
Test: video port valid and sendrecv scenario verified
Change-Id: I185f5d8773375be35d74fd8efbb48f6c5b545800