Do not restart IMS sevice when SIM removed
When SIM is inserted and removed, the IMS service is restarted because
the ImsStack recognizes that the carrier is changed.
This change is not to restart the IMS service even if the SIM is removed
to guarantee the IMS de-registration if the IMS is already registered.
When the same or different SIM is inserted, the IMS service will be
restarted with a new  carrier configuration.
Regarding to the configuration update on SIM ABSENT, this will be
changed in b/266032247.

Bug: 265609880
Test: manual (verify that service is not restarted when removing SIM
card)

Change-Id: I2d86bb9075da181fd69073edda9a8eb9ea8bb3a9