(Mtc) Silent re-dial the emergency call by E2 timer
It is to add implementation to handle the silent emergency re-dial with
anonymous INVITE case. The cases are required by the T-Mobile US
specification that E2 timer is expired or UE receives 4xx/5xx/6xx
response. E2 timer is for waiting 1xx/18x/2xx response to INVITE
request. And changed to retain the E3 timer if the emergency call is
re-dialed by E2 timer.

Bug: b/267157594
Test: verified the case for no receiving 100 Trying response to INVITE request (E2 timer expiry)
Test: verified case for receiving 4xx/5xx/6xx response to INVITE request without 100 Trying response (E2 timer is active)
Test: atest passed

Change-Id: I13f817725a1372adb5e6eaff758f45f60d6891b3