Hang in Hanging up screen is fixed
Initial IMediaConnectionObserver instance was held by ImsMediaManagerCallback. When ImsMediaService disconnected callback is received, events were posted to initial instance even for later calls resulting in audioSessionClosed event not reached  to respective MtcMediaSession instance. This is rectified by holding it in static, so that only latest instance will be used.
And quit HandlerThread when ImsMediaManager is released.

Bug: 266411958
Test: verified with stress test
      atest ImsStackJavaTests
Change-Id: I0436ebb3f2af458460b22dca41297fb7992707c9