(Ut) Removing PDN connection handling logic for other SIM
This is CL to remove PDN control logic for other SIM. It disconnects
XCAP APN for other SIM before trying to connect XCAP APN with current
SIM. But it's not veryfied with new HAL interface that includes
start/stop IMS traffic. So, it will be implemented when there are any
issues.

Bug: 205083701
Test: atest ImsStackJavaTests
Change-Id: Idbd2165d5ecd5cc10de922844c25533c57245177