Fix a memory violation in SipMessageBodyPart
In the process of processing the forked response to the INVITE response,
the SIP message body is cloned, and a memory violation occurred due to
an incorrect memory access during this process.
This CL is to fix the memory access violation by re-parsing the content
header and buffer when cloning the SIP message body.

Bug: 266140073
Bug: 269365967
Test: manual (hwasan enabled, IMS MO call and receive forked response)
Test: atest ImsStackNativeTests
Change-Id: Ic83c508c1a97e73d390e00ddcc953104abe8ab6f