(IMS_MEDIA) use current direction if no prevDirection
In SRVCC cases, if the DUT receives a SRVCC_CANCEL or SRVCC_FAIL message, it will send a ModifySession message including the previous cached direction. However, the previous direction is only cached when the DUT receives a SRVCC_START message. If the DUT does not receive a SRVCC_START message before receiving a SRVCC_CANCEL or SRVCC_FAIL message, the previous direction will not be cached. In this case, the DUT will send a ModifySession message including a default value (MEDIA_DIRECTION_INVALID).
DUT will cache the previous direction even if it didn’t receive a SRVCC_START message first. This way, when DUT gets only a SRVCC_CANCEL or SRVCC_FAIL message, it will still send a ModifySession message with the previous direction.

Bug: 263930262
Test: passed L_IMS_VOLTE_6826_9
      atest ImsStackNativeEnablerMediaTests
Change-Id: I68039745f30074fb80fb5f7f4cea70eb0a2a92f8