(Mtc) Check emergency call via CallInfo in CsCallBlockRule
It is the change to check if the current call is emergency or not via
CallInfo. Previously, it checked with the service type of the call if it
is emergency. But there can be an emergency call with normal service
type, so this change is needed.

Bug: b/269705243
Test: atest ImsStackNativeEnablerMtcTests with hwasan
Change-Id: Ic8093fb8408fda68d5973a825b3f2b4322307381
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:9ee0543153a6acf26e6e38d0eadfebee12be7939)
Merged-In: Ic8093fb8408fda68d5973a825b3f2b4322307381