(Mtc) Fix to close and destroy the media session
It is not to use close and destroy the media session separately.
Previously, MTC module called MediaSession::Terminate() to close the
media session and MediaManager::DestroySession() to remove the session.
Terminate() can be invoked in case of call termination and silent redial call.
After closing the media session, there is no scenario to use the exist
one again. So MTC and Media module decided to remove Terminate()
interface and close the media session through DestroySession().

Bug: b/270674517
Test: atest passed
Test: simulated test the case of silent redial and call termination
Change-Id: I64d8af8bc870c4b3d5bff88ed10ab459e161ce46