Set a proper refresher parameter when no Session-Expires header
If the incoming request contains a Supported header field with a value
"timer", but does not contain a Session-Expires header, it means that
the UAS is indicating support for timers but is not requesting one. UAS
may request a session timer in the 2xx response by including a
Session-Expires header field. When the Session-Expires header field is
included, the "refresher" parameter will be set to "uac" according to
the standard specification(GSMA IR.92).

Bug: 270932653
Test: manual (verify refresher parameter when INVITE received w/o
Session-Expires header)

Change-Id: I40648b9e60ae82ce3f722cd0e3503282747b05a2