(Mtc) Remove the configuration to check verstat capability
This change is to remove the asset field for checking the network
capability of calling number verification(verstat) from the result of
IMS registration.
It is not mandatory to check the capability of verstat from the
requirements, so this field is not necessary anymore.

Bug: b/235749454
Test: atest passed
Change-Id: I3970f477c506e34a4b98279e902fcaf72fda1644
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:0418680d16a491926afe3673edb90688fbc0d936)
Merged-In: I3970f477c506e34a4b98279e902fcaf72fda1644