(IMS_MEDIA) Remove definition REPORT_CLOSE_SESSION
Remove definition of REPORT_CLOSE_SESSION which is not used.
Removed handling REPORT_CLOSE_SESSION in MtcMediaManager#MediaSession_Notify.(ag/21853754)
Because it doesn't have to receive the report type related to close media session anymore.
MediaSession will be closed when MTC module invokes DestroySession for Media enabler.
So the result of closing the media session is not useful to MTC module.

Bug: b/272395523
Test: build passed
Change-Id: Ie7b4d064c765f2ec0b1ab3a58834c29e4eae400e