(config) Disable TRANSPORT_ERROR_REPORT_ON_TXN option
Basically, SIP transactions are managed independently of transport. In
some cases, certain carriers require that the lifecycle of SIP
transactions be managed the same as the transport lifecycle.
Currently, the configuration for applying this requirement is always
enabled, so this change is to disable it so that SIP transactions are
managed independently of transport.

Bug: 273393099
Bug: 264133707
Test: atest ImsStackNativeTests
Change-Id: I61b1867f85b176d026b58934260354733408e1ad