(Ut) Initializing SscConfig while binding the CallApp
This is a CL to initialize SscConfig when CallApp is binded. After SIM
SWAP, ImsServiceManager unbind and bind CallApp if the CallApp is
already binded. In this case, SscServiceImpl previously created is used
again. SscConfig was initialized only When SscServiceImpl is created.
So, it's changed to initialize SscConfig when SscServiceImpl is started.

Bug: 274408140
Test: atest ImsStackJavaTests
Test: No sim to TMUS SIM swap test
Change-Id: Idf5c8856a2e86cae4ffe53f3975b0f7647dbe53c