(core) Restore server connection early during re-registration
If the server socket is closed due to the handover (IWLAN/LTE) when IMS
is in the REGISTERED state, IMS re-registration proceeds normally, but
there was a problem that the server socket recovery did not proceed
normally after re-registration was completed. (After re-registration is
complete, the server socket is cleaned up, and a bind failure occurs
even if the socket is created after a few seconds.)
To avoid this situation(bind failure), the registration component will
recover the server socket early if the server socket recovery is
required during IMS re-registration.

Bug: 269562225
Test: manual (verify server connection restoration while re-registering)
Change-Id: I18cf6950b10b3da8f1a68d1d7856d90e3cd1e04e