assertParcelEquals -= call to recycle
Many of the clients which call this method, such as
testNotifyQosInfo and many others, assume that they own
the Parcel objects that are being sent here.

So, I'm removing these, because they will cause the Parcel
objects to be double-freed, and this can lead to arbitrary
memory corruption in these tests.

When you use Parcel, always use it like this:

    Parcel p = Parcel.obtain();
    try {
        ... use p ...
    } finally {
        p.recycle();
    }

This way, it is always obtained and recycled exactly once.

Bug: 243768096
Test: N/A
Change-Id: I16331bf333fdbe9814a2bd94fbd0279697e2c90a