Fix a memory leak of SipTxnKey in INVITE client transaction
In case of INVITE client transaction, we need to update the to-tag when
receiving 18x or 200 response to distinguish the forked responses. To do
this, SIP stack returns a new SipTxnKey when the INVITE response is
received, and SipClientTransactionState updates the new transaction key
containing the to-tag. The memory leak was detected in this procedure.
This change removes the increment of the reference count of SipTxnKey in
this situation.

Bug: 261937752
Test: verify IMS MO/MT call on target device
Change-Id: I3fb1b1e4333fba7f6d7a53319361eecc71621a7e