(Mtc) Move update reject code check logic from GII to MTC
ImsVideoCallSession is checking the configuration value and choose the
API of MTC to call.
1. if reject code 200 : accept
2. if reject code other than 200 : reject
However, if precondition call flow is used by native, there is no way to
send 200 OK with SDP. So, the configuration must be checked by native
MTC. This commit removes the logic in GII and adds a new checking logic
in MTC side.

Bug: 274840426
Test: verified 4 combinations (reject code X call flow type)
Test: atest ImsStackNativeEnablerMtcTests, ImsStackJavaTests
Change-Id: I8a29bfa79681a91e70f6a68f72b56113465cb662