(Mtc) Do not update CallType if the new one is same as the original one.
MtcSession internally updates the member CallType when a request or
response is received. When the CallType is updated, the original
CallType is also updated by the current one. However, the updating logic
doesn't check the equality of the newly updating type so the original
CallType loses its own meaning when the SIP message below is received.
- Provisional response
- Early UPDATE
- 200 OK for Early UPDATE
This commit changes MtcSession not to update the CallType if the newly
updating CallType is same as the current one and the SIP message is not
the initiaing session modification. (eg. re-INVITE)

Bug: 274840426
Test: verified using Add Video Call flow test code
Test: atest ImsStackNativeEnablerMtcTests
Change-Id: Ib0824e43f5c37b2cffecdc134dd27e8b56cf1470