(imscore) Support SDP re-offer when PRACK sent
Currently, PRACK request can include the SDP answer when the UE sends
the INVITE request without the SDP body and the network sends 183
response with the SDP body. However, if the UE needed to send SDP as an
SDP re-offer in the PRACK request, it was not supported.
This change allows the enabler to send an SDP re-offer in the PRACK
request when the enabler needs to include the SDP body according to the
standard specification or carrier requirements.

Bug: 278822559
Test: manual (verify PRACK request includes SDP re-offer)
Change-Id: I0c15605b6fb4b615a942c16da30616df47b75867