(Mtc) Just one pending operation is processed when there are pending operation
If a pending operation occurred more than one almost at the same time and the topmost pending operation is not changing the state of the call,
remained pending operations are alive during the call(because runs a topmost pending operation when enters at the EstablishState) and might do something wrong when enters at the EstablishState next time.
Also next time occurred pending operations are not running correctly.

Bug: b/277672581
Test: ImsStackNativeEnablerMtcTests
Test: Simulator test

Change-Id: I9ce8660a9ac65782fcb2b85038e3de1d5ffc047b