(imscore) Fix SDP answer when unchanged SDP re-offer receives
SIP engine handles the SDP offer/answer based on the negotiated
(selected) SDP. However, RFC3264 specifies that the SDP offer/answer is
based on the last SDP provided to the other party. Therefore, if UE
receives the re-INVITE request with the negotiated codec from the other
party, it needs to send the SDP answer with the increase of the SDP
session version because UE's last SDP provided to the other party was a
full capability of the current session.
This CL changes the following cases:
 - SDP answerer: if the SDP answer and the last SDP provided is
   different, then increase the SDP session version.
 - SDP offerer: if the SDP re-offer (session refresh by re-INVITE) and
   the last SDP provided is different, then increase the SDP session
   version.

Bug: 248985767
Test: manual (verify SDP offer/answer for MO/MT call setup,
modification)

Change-Id: Iaa37e0ed7fc0d77276da5b82e8ea84a7a48fc383