Refactor DcFactory class
This is to change the management of the data connection related
components in the DcFactory class. In the previous implementation, each
component could be accessed with the type identifier, but in the changed
implementation, it could be accessed with the interface class type of
each component.
This change contains the followings:
 - remove the predefined types that identify each component
 - change API surface with an interface class type
 - update the caller side that is related to this change

Bug: 286305568
Test: atest ImsStackJavaTests
Test: manual (verify MO/MT call on device)
Change-Id: I7ba825c81bdfdf13e58e8f8cd69ba40267e2f711