(imscore) Add SDP reader interface for 488-INVITE
A UAS rejecting an offer contained in an INVITE may return a 488 (Not
Acceptable Here) or 606 (Not Acceptable) response. Such a response may
contain the message body containing a description of media capabilities.
To handle this message body part properly in the enabler layer, the
ISession provides an interface so that the enablers can access the SDP
body part with the same IMediaDescriptor as a normal SDP exchange.

New interface added:
 - ISession#GetRemoteMediaCapabilities() can be invoked when 488 or 606
   response for the INVITE request is received and returns ISdpReader.
 - ISdpReader contains an ISessionDescriptor and one or multiple
   IMediaDescriptors.

Bug: 277531866
Test: atest ImsStackNativeTests
Test: manual (verify 488-INVITE response with SDP)
Test: manual (verify IMS reg/call on device)
Change-Id: I61886c2c6e31c00e7acf2014760e62411557eb47