Fix crash issue when restarting ImsStack process
This crash was caused by comparing the slot index provided and the size
of the object container when the service started in slot1 before the
service started in slot0.
To avoid this situation, the getter method will not check the given slot
index since the container is a hash map and returns a null object if the
given slot index does not exist.

Bug: 292768306
Test: atest ImsStackJavaTests
Test: manual (verify IMS reg on device)
Change-Id: If5e884586ba15e7023cdda2bc0b5cd633c1d453f