(sipcore) Enable Session-ID header for REGISTER/INVITE only
The Session-ID header may be included in all the SIP messages, but as
per a RFC 7329, it describes the specific behaviors for REGISTER and
INVITE request including in-dialog messages.
So we will enable Session-ID header for REGISTER and INVITE requests at
this time and other messages may be enabled if necessary in the future.

Bug: 297445811
Test: manual (verify Session-ID header is present in INVITE and its
related dialog messages)
Test: manual (verify Session-ID header is not present except for
REGISTER and INVITE)

Change-Id: I062e93eeaf5543157adc2054ad6c0023eec9b837