add null check and media state check.
java exception by mLocalAddress was null.
local member variable like mLocalAddress or mRtpSocketList are set when the handleAudioOpenSession is invoked, but in case of ImsMedia is not connected at this time, openSession process is aborted and those variables could remain as null.
so added null check and media state check for handleAudioQos.

Bug: 317360160
Test: mm and verify this case and MO/MT
      atest ImsMediaJavaUnitTests
Change-Id: I58befe165028e56177974a026b52300146eabdd2