Fix incorrect display of hidden menu
When entering ImsStack's hidden menu, there was a problem where the
first menu item was not visible due to the title bar or was located in
the position of the system bar.
This change is to position the first menu item below the title/system
bar so that the menu can be used normally. This is also applied to
configure the menu according to the device's theme style.

Bug: 332991154
Test: atest ImsStackJavaTests
Change-Id: I63260319ddcb7a451e854cfb253b64c736457fb8